refactor: extract PangoLineageCounts
issue: #39
